Refactor notification page
