+++ src/pages/notification/index.js
@@ -1,0 +1,12 @@
+import NotifIcon from "@mui/icons-material/PriorityHigh";
+
+import NotifList   from "./NotifList";
+import NotifEdit   from "./NotifEdit";
+import NotifCreate from "./NotifCreate";
+
+export default {
+    create: NotifCreate,
+    icon:   NotifIcon,
+    edit:   NotifEdit,
+    list:   NotifList
+};
